+++ src/components/TextInput/index.test.js
@@ -1,17 +1,9 @@
 import React from 'react';
-import { shallow, mount } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import TextInput, { throttleInput } from '.';
+import TextInput from '.';
 
 describe('TextInput component', () => {
-
-  it('fires throttleFunctions', () => {
-    const throttleFn = jest.fn()
-    const wrapper = mount(<TextInput onThrottledChange={throttleFn} label="label" />);
-    const input = wrapper.find('input')
-    input.simulate('change', { target: { value: 'aaa' } });
-    expect(throttleFn).toHaveBeenCalled();
-  });
 
   it('matches snapshot (default)', () => {
     const wrapper = shallow(<TextInput label="label" />);
@@ -26,7 +18,6 @@ describe('TextInput component', () => {
         IconLeft={() => {}}
         IconRight={() => {}}
         onChange={() => {}}
-        onThrottledChange={() => {}}
         before="before"
         after="after"
         type="number"
@@ -36,35 +27,6 @@ describe('TextInput component', () => {
       />
     );
     expect(wrapper).toMatchSnapshot();
-  });
-
-});
-
-describe('throttleInput', () => {
-  let throttleFn;
-  let debounceFn
-
-  beforeEach(() => {
-    throttleFn = jest.fn();
-    debounceFn = jest.fn();
-  })
-
-  it('calls throttleFn when less than 5 characters', () => {
-    throttleInput('aaa', throttleFn, debounceFn)
-    expect(throttleFn).toHaveBeenCalled();
-    expect(debounceFn).not.toHaveBeenCalled();
-  });
-
-  it('calls throttleFn when a space is the last character', () => {
-    throttleInput('aaaaaaaaaaa ', throttleFn, debounceFn)
-    expect(throttleFn).toHaveBeenCalled();
-    expect(debounceFn).not.toHaveBeenCalled();
-  });
-
-  it('calls debounceFn with 5 characters or more', () => {
-    throttleInput('aaaaa', throttleFn, debounceFn)
-    expect(throttleFn).not.toHaveBeenCalled();
-    expect(debounceFn).toHaveBeenCalled();
   });
 
 });