+++ src/components/Tooltip/index.jsx
@@ -1,43 +1,34 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import FDS from 'lib/dictionary/js/styleConstants';
-import Popover, { VALID_POSITIONS } from 'components/Popover';
-
-export const DEFAULT_WIDTH = '240px';
-export const DELAY_MS = 350;
-export const DEFAULT_POSITION = 'bottom';
-const DEFAULT_TEXT_ALIGN = 'center';
-export const VALID_TEXT_ALIGN = ['left', 'center', 'right'];
+import Tippy from '@tippyjs/react';
 
 const Tooltip = ({
   trigger,
   message,
-  maxWidth,
-  position = DEFAULT_POSITION,
-  textAlign = DEFAULT_TEXT_ALIGN,
+  position = 'bottom',
+  maxWidth = 240,
+  textAlign = 'center',
 }) => (
-  <Popover
-    interactionMode="hover"
-    trigger={<span style={{ cursor: 'help' }}>{trigger}</span>}
-    distance={parseInt(FDS.SPACE_HALF, 10)}
-    position={position}
-    alignment="center"
-    transitionName="GrowFast"
-    delay={DELAY_MS}
+  <Tippy
+    arrow={false}
+    theme="tooltip"
+    delay={[350, 0]}
+    maxWidth={maxWidth}
+    content={message}
+    placement={position}
+    zIndex={FDS.ZINDEX_MODAL}
+    moveTransition="transform 0.2s ease-out"
+    appendTo={() => document.body}
+    className={`bgColor--charcoal inverted align--${textAlign} padding--y--s  padding--x elevation--2 rounded--all fontSize--s fontWeight--bold`}
   >
-    <div
-      role="tooltip"
-      style={{ maxWidth: maxWidth ? `${maxWidth}px` : DEFAULT_WIDTH }}
-      className={`bgColor--charcoal inverted align--${textAlign} padding--top--half padding--bottom--half padding--left padding--right elevation--2 rounded--all fontSize--s fontWeight--bold`}
-    >
-      {message}
-    </div>
-  </Popover>
+    {trigger}
+  </Tippy>
 );
 
 Tooltip.propTypes = {
   /** JSX - Acts as a positioning reference for the popover and triggers active state */
-  trigger: PropTypes.oneOfType([PropTypes.node, PropTypes.element]).isRequired,
+  trigger: PropTypes.oneOfType([PropTypes.node, PropTypes.element]),
 
   /** Content of the tooltip */
   message: PropTypes.string.isRequired,
@@ -45,14 +36,10 @@ Tooltip.propTypes = {
   /** Maximum width of tooltip */
   maxWidth: PropTypes.number,
 
-  /**
-   * Position preference of tooltip.
-   * `top` for example, will place the tooltip above the trigger.
-   */
-  position: PropTypes.oneOf(VALID_POSITIONS),
-
+  /** Position preference of tooltip. */
+  position: PropTypes.oneOf(['auto', 'top', 'right', 'bottom', 'left']),
   /** Text align for the message */
-  textAlign: PropTypes.oneOf(VALID_TEXT_ALIGN),
+  textAlign: PropTypes.oneOf(['left', 'center', 'right']),
 };
 
 export default Tooltip;