+++ src/components/Tooltip/index.stories.js
@@ -5,8 +5,16 @@ const Template = (args) => <Tooltip {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {
-  trigger: <span>ℹ️ Hover Here!</span>,
-  message: 'message',
+  trigger: <button>ℹ️ Hover Here!</button>,
+  message: 'Hello world',
+};
+
+export const maxWidth = Template.bind({});
+maxWidth.args = {
+  maxWidth: 600,
+  trigger: <button>ℹ️ Hover Here!</button>,
+  message:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
 };
 
 export default {